code/requirements.txt
openai
python-dotenv
pydantic
requests
langchain-community==0.2.7
multion
supabase